teset ing

+++ cart.php
@@ -1,14 +1,9 @@
-<?php 
-    session_start();
-    if (!$_SESSION['login']) {
-        header('location: login.php');
-        exit;
-    }
+<?php
 
-    include 'connect.php';
+include 'connect.php';
 
-    // Mengambil data dari table carts
-    $queryShowCarts = mysqli_query($conn, "SELECT * FROM carts ORDER BY name_product");
+// Mengambil data dari table carts
+$queryShowCarts = mysqli_query($conn, "SELECT * FROM carts ORDER BY name_product");
 ?>
 
 <!DOCTYPE html>
@@ -20,8 +15,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1" />
 
     <!-- Bootstrap CSS -->
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous" />
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-1BmE4kWBq78iYhFldvKuhfTAU6auU8tT94WrHftjDbrCEXSU1oBoqyl2QvZ6jIW3" crossorigin="anonymous" />
 
     <!-- Bootsrap Icon -->
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.7.2/font/bootstrap-icons.css" />
@@ -85,12 +79,12 @@
                 </thead>
                 <tbody>
                     <?php
-                        $no = 1;
-                        $total = 0;                    
+                    $no = 1;
+                    $total = 0;
                     ?>
                     <?php while ($data = mysqli_fetch_assoc($queryShowCarts)) : ?>
                         <tr>
-                            <th scope="row"><?= $no++.'.' ?></th>
+                            <th scope="row"><?= $no++ . '.' ?></th>
 
                             <td><?= $data['name_product'] ?></td>
 
@@ -101,7 +95,7 @@
                             </td>
 
                             <td class="text-center">Rp<?= number_format($data['total_price'], 0, '', '.') ?></td>
-                            
+
                             <td class="text-center">
                                 <img src="img/product-img/<?= $data['img_product'] ?>" width="100px">
                             </td>
@@ -121,24 +115,22 @@
             </table>
             <div class="d-flex justify-content-between">
                 <h5 class="text-white fw-bold">Order Total : <span class="colorAcsent fs-4">Rp<?= number_format($total, 0, '', '.') ?></span></h6>
-                <div>
-                    <a class="btn btn-danger me-1" href="process.php?delete=all"
-                        onclick="return confirm('Yakin ingin mengosongkan cart?'); ">Empty Cart</a>
-                    <a class="btn btn-primary" href="checkout.php">Checkout</a>
-                </div>
+                    <div>
+                        <a class="btn btn-danger me-1" href="process.php?delete=all" onclick="return confirm('Yakin ingin mengosongkan cart?'); ">Empty Cart</a>
+                        <a class="btn btn-primary" href="checkout.php">Checkout</a>
+                    </div>
             </div>
         </form>
 
-        <?php 
-            
+        <?php
+
         ?>
         </h1>
     </div>
     <!-- AKhir table cart -->
 
 
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js"
-        integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous">
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous">
     </script>
 </body>
